Add icons to side nav quick links, Add styling to space icons with text , Move farmsdropdwon file to components
