--- FRONT-END/LeafLife/src/pages/gardendata.js
@@ -1,18 +1,0 @@
-// gardensData.js
-
-const gardensData = [
-    {
-    id: 1,
-    name: "North Garden",
-    },
-    {
-    id: 2,
-    name: "South Garden",
-    },
-    {
-    id: 3,
-    name: "East Garden",
-    },
-];
-
-export default gardensData;
